Move all error interceptors in one file
