+++ pkg/grpc/errors/interceptor.go
@@ -1,4 +1,4 @@
-package grpccode
+package errors
 
 import (
 	"context"
@@ -11,6 +11,26 @@ import (
 	"google.golang.org/grpc/codes"
 	"google.golang.org/grpc/status"
 )
+
+// PanicOnInvariantViolationUnaryInterceptor panics on ErrInvariantViolation.
+// Note: this interceptor should ONLY be used in dev builds.
+func PanicOnInvariantViolationUnaryInterceptor(ctx context.Context, req interface{}, _ *grpc.UnaryServerInfo, handler grpc.UnaryHandler) (interface{}, error) {
+	resp, err := handler(ctx, req)
+	if errors.Is(err, errorhelpers.ErrInvariantViolation) {
+		panic(err)
+	}
+	return resp, err
+}
+
+// PanicOnInvariantViolationStreamInterceptor panics on ErrInvariantViolation.
+// Note: this interceptor should ONLY be used in dev builds.
+func PanicOnInvariantViolationStreamInterceptor(srv interface{}, ss grpc.ServerStream, _ *grpc.StreamServerInfo, handler grpc.StreamHandler) error {
+	err := handler(srv, ss)
+	if errors.Is(err, errorhelpers.ErrInvariantViolation) {
+		panic(err)
+	}
+	return err
+}
 
 // ErrorToGrpcCodeInterceptor translates common errors defined in errorhelpers to GRPC codes.
 func ErrorToGrpcCodeInterceptor(ctx context.Context, req interface{}, info *grpc.UnaryServerInfo, handler grpc.UnaryHandler) (interface{}, error) {
